Lab2/pyproject.toml
[tool.poetry]
name = "lab2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8, <3.10"
numpy = "^1.23.4"
Pillow = "^9.3.0"
rich = "^12.6.0"
tqdm = "^4.64.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
